- Implemented ShotgunUser proxy class. - Proxy class uses properties instead of methods.
