+++ tests/shotgun_authentication_tests/test_shotgun_wrapper.py
@@ -14,14 +14,12 @@ from mock import patch
 from tank_test.tank_test_base import *
 
 from tank_vendor.shotgun_api3 import AuthenticationFault
-from tank_vendor.shotgun_authentication import user, ShotgunAuthenticationError
+from tank_vendor.shotgun_authentication import user_impl, ShotgunAuthenticationError
 
 
 class ShotgunWrapperTests(TankTestBase):
     """
-    Tests the user module. Note that because how caching the session information is still
-    very much in flux, we will not be unit testing cache_session_info, get_session_data and
-    delete_session_data for now, since they have complicated to test and would simply slow us down.
+    Tests the user_impl module.
     """
 
     @patch("tank_vendor.shotgun_api3.Shotgun._call_rpc")
@@ -36,7 +34,7 @@ class ShotgunWrapperTests(TankTestBase):
         mocked_result = {"entities": [1, 2, 3]}
         _call_rpc_mock.side_effect = AuthenticationFault()
 
-        sg_user = user.SessionUser("https://host.shotgunstudio.com", "login", "session", "proxy")
+        user = user_impl.SessionUser("https://host.shotgunstudio.com", "login", "session", "proxy")
         # Directly call _call_rpc. We should be invoking the derived class here, which will
         # then invoke the base class which is in fact our mock class so it should throw once and then
         # succeed.
@@ -47,7 +45,7 @@ class ShotgunWrapperTests(TankTestBase):
             _call_rpc_mock.side_effect = [mocked_result]
 
         renew_session_mock.side_effect = renew_session_side_effect
-        test_result = sg_user.create_sg_connection()._call_rpc()
+        test_result = user.create_sg_connection()._call_rpc()
 
         # Make sure we tried to renew the sesion
         self.assertTrue(renew_session_mock.called)
@@ -68,9 +66,9 @@ class ShotgunWrapperTests(TankTestBase):
         _call_rpc_mock.side_effect = AuthenticationFault("This is coming from the _call_rpc_mock.")
         renew_session_mock.side_effect = ShotgunAuthenticationError("This is coming from renew_session_mock.")
 
-        sg_user = user.SessionUser("https://host.shotgunstudio.com", "login", "session", "proxy")
+        user = user_impl.SessionUser("https://host.shotgunstudio.com", "login", "session", "proxy")
         with self.assertRaisesRegexp(ShotgunAuthenticationError, "This is coming from renew_session_mock."):
-            sg_user.create_sg_connection()._call_rpc()
+            user.create_sg_connection()._call_rpc()
 
         # Make sure we tried to renew the sesion
         self.assertTrue(renew_session_mock.called)
